[refactor] Ladder, HorizontalLine 테스트서 동시에 사용되는 패턴 클래스로 분리
