+++ src/test/java/ladder/domain/HorizontalLineTest.java
@@ -1,6 +1,5 @@
 package ladder.domain;
 
-import ladder.ladderexceptions.InvalidLayoutException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,7 +10,6 @@ import java.util.Arrays;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class HorizontalLineTest {
 
@@ -21,7 +19,7 @@ class HorizontalLineTest {
                         new Boolean[]{false, false, false, false, false}),
                 Arguments.of((HowToConnect) leftSideStatus -> !leftSideStatus,
                         new Boolean[]{true, false, true, false, true}),
-                Arguments.of(new SampleConcretePattern(),
+                Arguments.of(new SampleShufflePattern(),
                         new Boolean[]{false, true, false, true, false})
         );
     }
@@ -54,32 +52,5 @@ class HorizontalLineTest {
 
         assertThat(sampleLine)
                 .isEqualTo(HorizontalLine.ofLines(Arrays.asList(expected)));
-    }
-
-    @Test
-    @DisplayName("허용되지 않는 배열(연속 True)에 대한 예외처리 테스트")
-    void testInvalidCondition() {
-        Boolean[] input = {true, false, true, true};
-
-        assertThatThrownBy(
-                () -> HorizontalLine.ofLines(Arrays.asList(input))
-        ).isInstanceOf(InvalidLayoutException.class);
-    }
-
-    static class SampleConcretePattern implements HowToConnect {
-        private int idx = 0;
-
-        public SampleConcretePattern() {
-        }
-
-        @Override
-        public boolean needToConnect(Boolean leftSideCondition) {
-            idx += 1;
-            return isEven(idx);
-        }
-
-        private boolean isEven(int index) {
-            return index % 2 == 0;
-        }
     }
 }
